refactor thumbnail endpoints
test thumbnail generation locally to assert file actually created
