+++ backend/app/main.py
@@ -7,15 +7,13 @@ from fastapi import (
     BackgroundTasks,
     Body,
     Form,
-    Security,
 )
-from fastapi.responses import RedirectResponse
 from typing import Annotated
 import botocore.exceptions
 from sqlalchemy.exc import MultipleResultsFound, NoResultFound, DataError
 from fastapi.responses import FileResponse
-from sqlalchemy import text, update
-from sqlmodel import Session, String, select, true
+from sqlalchemy import text
+from sqlmodel import Session, String, select, true, update
 from sqlalchemy.sql.functions import coalesce
 from starlette.middleware.cors import CORSMiddleware
 from sqlalchemy.dialects.postgresql import insert
@@ -25,15 +23,14 @@ from sqlmodel import ARRAY, INT
 from datetime import datetime, UTC, timedelta
 import sentry_sdk
 from fastapi_utils.tasks import repeat_every
-from app.core.db import get_session
+from app.core.db import get_session, get_document
 from app.core.config import settings
-from app.core.security import auth, TokenScope
-from app.core.thumbnails import generate_thumbnail, thumbnail_exists
 from app.utils import hash_password, verify_password
 import jwt
 from uuid import uuid4
 import app.contiguity.main as contiguity
 import app.cms.main as cms
+import app.thumbnails.main as thumbnails
 from networkx import Graph, connected_components
 from app.models import (
     Assignments,
@@ -60,7 +57,8 @@ from app.models import (
 from pydantic_geojson import PolygonModel
 from pydantic_geojson._base import Coordinates
 from sqlalchemy.sql import func
-from app.utils import remove_file, _cleanup_expired_locks, _remove_all_locks
+from app.utils import _cleanup_expired_locks, _remove_all_locks
+from app.io import remove_file
 from app.exports import (
     get_export_sql_method,
     DocumentExportType,
@@ -96,8 +94,9 @@ async def lifespan(app: FastAPI):
 
 app = FastAPI(lifespan=lifespan)
 
-# Include the CMS router
+# Module routers
 app.include_router(cms.router)
+app.include_router(thumbnails.router)
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.INFO)
@@ -327,28 +326,12 @@ async def update_assignments(
     return {"assignments_upserted": len(data.assignments), "updated_at": updated_at}
 
 
-def _get_document(
-    document_id: str, session: Session = Depends(get_session)
-) -> Document:
-    try:
-        document = session.exec(
-            select(Document).filter(Document.document_id == document_id)  # pyright: ignore
-        ).one()
-    except NoResultFound:
-        raise HTTPException(status_code=404, detail="Document not found")
-    except Exception as e:
-        logger.error(f"Error loading document: {str(e)}")
-        raise HTTPException(status_code=500, detail="Error loading document")
-
-    return document
-
-
 @app.patch(
     "/api/update_assignments/{document_id}/shatter_parents",
     response_model=ShatterResult,
 )
 async def shatter_parent(
-    document: Annotated[Document, Depends(_get_document)],
+    document: Annotated[Document, Depends(get_document)],
     data: GEOIDS,
     session: Session = Depends(get_session),
 ):
@@ -383,7 +366,7 @@ async def shatter_parent(
     response_model=GEOIDSResponse,
 )
 async def unshatter_parent(
-    document: Annotated[Document, Depends(_get_document)],
+    document: Annotated[Document, Depends(get_document)],
     data: AssignedGEOIDS,
     session: Session = Depends(get_session),
 ):
@@ -423,7 +406,7 @@ async def unshatter_parent(
     "/api/update_assignments/{document_id}/reset", status_code=status.HTTP_200_OK
 )
 async def reset_map(
-    document: Annotated[Document, Depends(_get_document)],
+    document: Annotated[Document, Depends(get_document)],
     session: Session = Depends(get_session),
 ):
     partition_name = f'"document.assignments_{document.document_id}"'
@@ -494,7 +477,7 @@ async def update_colors(
 # called by getAssignments in apiHandlers.ts
 @app.get("/api/get_assignments/{document_id}", response_model=list[AssignmentsResponse])
 async def get_assignments(
-    document: Annotated[Document, Depends(_get_document)],
+    document: Annotated[Document, Depends(get_document)],
     session: Session = Depends(get_session),
 ):
     stmt = (
@@ -504,13 +487,14 @@ async def get_assignments(
             Assignments.document_id,
             ParentChildEdges.parent_path,
         )
-        .join(Document, Assignments.document_id == Document.document_id)
+        .join(Document, onclause=Assignments.document_id == Document.document_id)
         .join(
-            DistrictrMap, Document.districtr_map_slug == DistrictrMap.districtr_map_slug
+            DistrictrMap,
+            onclause=Document.districtr_map_slug == DistrictrMap.districtr_map_slug,
         )
         .outerjoin(
             ParentChildEdges,
-            (Assignments.geo_id == ParentChildEdges.child_path)
+            onclause=(Assignments.geo_id == ParentChildEdges.child_path)
             & (ParentChildEdges.districtr_map == DistrictrMap.uuid),
         )
         .where(Assignments.document_id == document.document_id)
@@ -655,7 +639,7 @@ async def get_document_status(
 
 @app.get("/api/document/{document_id}/unassigned", response_model=BBoxGeoJSONs)
 async def get_unassigned_geoids(
-    document: Annotated[Document, Depends(_get_document)],
+    document: Annotated[Document, Depends(get_document)],
     exclude_ids: list[str] = Query(default=[]),
     session: Session = Depends(get_session),
 ):
@@ -930,7 +914,7 @@ async def get_connected_component_bboxes(
 
 @app.put("/api/document/{document_id}/metadata", status_code=status.HTTP_200_OK)
 async def update_districtrmap_metadata(
-    document: Annotated[Document, Depends(_get_document)],
+    document: Annotated[Document, Depends(get_document)],
     metadata: dict = Body(...),
     session: Session = Depends(get_session),
 ):
@@ -961,7 +945,7 @@ async def get_projects(
 ):
     gerrydb_views = session.exec(
         select(DistrictrMap)
-        .filter(DistrictrMap.visible == true())  # pyright: ignore
+        .where(DistrictrMap.visible == true())
         .order_by(DistrictrMap.created_at.asc())  # pyright: ignore
         .offset(offset)
         .limit(limit)
@@ -969,47 +953,9 @@ async def get_projects(
     return gerrydb_views
 
 
-@app.get("/api/document/{document_id}/thumbnail", status_code=status.HTTP_200_OK)
-async def get_thumbnail(*, document_id: str, session: Session = Depends(get_session)):
-    try:
-        if thumbnail_exists(document_id):
-            return RedirectResponse(
-                url=f"{settings.CDN_URL}/thumbnails/{document_id}.png"
-            )
-        else:
-            return RedirectResponse(url="/home-megaphone.png")
-    except:
-        return RedirectResponse(url="/home-megaphone.png")
-
-
-@app.post("/api/document/{document_id}/thumbnail", status_code=status.HTTP_200_OK)
-async def make_thumbnail(
-    *,
-    document_id: str,
-    background_tasks: BackgroundTasks,
-    session: Session = Depends(get_session),
-    auth_result: dict = Security(auth.verify, scopes=[TokenScope.create_content]),
-):
-    try:
-        stmt = select(Document.document_id).filter(Document.document_id == document_id)
-        map = session.execute(stmt).first()
-    except DataError:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Document ID did not fit UUID format",
-        )
-    if map == None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Document not found",
-        )
-    background_tasks.add_task(generate_thumbnail, session=session, document_id=document_id)
-    return {"message": "Generating thumbnail in background task"}
-
-
 @app.post("/api/document/{document_id}/share")
 async def share_districtr_plan(
-    document: Annotated[Document, Depends(_get_document)],
+    document: Annotated[Document, Depends(get_document)],
     params: dict = Body(...),
     session: Session = Depends(get_session),
 ):
@@ -1152,7 +1098,7 @@ async def load_plan_from_share(
 
 @app.post("/api/document/{document_id}/checkout", status_code=status.HTTP_200_OK)
 async def checkout_plan(
-    document: Annotated[Document, Depends(_get_document)],
+    document: Annotated[Document, Depends(get_document)],
     params: dict = Body(...),
     session: Session = Depends(get_session),
 ):
@@ -1197,7 +1143,7 @@ async def checkout_plan(
 @app.get("/api/document/{document_id}/export", status_code=status.HTTP_200_OK)
 async def get_survey_results(
     *,
-    document: Annotated[Document, Depends(_get_document)],
+    document: Annotated[Document, Depends(get_document)],
     background_tasks: BackgroundTasks,
     format: str = "CSV",
     export_type: str = "ZoneAssignments",
